+++ src/components/core/Auth/Template.tsx
