+++ src/components/core/Homepage/CourseCard.tsx
